+++ backend-rust/src/transaction_event/mod.rs
@@ -3,6 +3,7 @@ use crate::{
     decoded_text::DecodedText,
     graphql_api::{ApiError, ApiResult},
     scalar_types::{BigInteger, Byte, DateTime, UnsignedLong},
+    transaction_event::transfers::TimestampedAmount,
 };
 use anyhow::Context;
 use async_graphql::{ComplexObject, Object, SimpleObject, Union};
@@ -304,11 +305,15 @@ pub fn events_from_summary(
                 vec![Event::TransferredWithSchedule(transfers::TransferredWithSchedule {
                     from_account_address: details.sender.into(),
                     to_account_address:   to.into(),
-                    total_amount:         amount
-                        .into_iter()
-                        .map(|(_, amount)| amount.micro_ccd())
-                        .sum::<u64>()
-                        .into(),
+                    amounts_schedule:     amount
+                        .iter()
+                        .map(|(timestamp, amount)| {
+                            Ok::<TimestampedAmount, anyhow::Error>(TimestampedAmount {
+                                timestamp: (*timestamp).try_into()?,
+                                amount:    UnsignedLong(amount.micro_ccd()),
+                            })
+                        })
+                        .collect::<Result<Vec<TimestampedAmount>, _>>()?,
                 })]
             }
             AccountTransactionEffects::TransferredWithScheduleAndMemo {
@@ -320,11 +325,15 @@ pub fn events_from_summary(
                     Event::TransferredWithSchedule(transfers::TransferredWithSchedule {
                         from_account_address: details.sender.into(),
                         to_account_address:   to.into(),
-                        total_amount:         amount
-                            .into_iter()
-                            .map(|(_, amount)| amount.micro_ccd())
-                            .sum::<u64>()
-                            .into(),
+                        amounts_schedule:     amount
+                            .iter()
+                            .map(|(timestamp, amount)| {
+                                Ok::<TimestampedAmount, anyhow::Error>(TimestampedAmount {
+                                    timestamp: (*timestamp).try_into()?,
+                                    amount:    UnsignedLong(amount.micro_ccd()),
+                                })
+                            })
+                            .collect::<Result<Vec<TimestampedAmount>, _>>()?,
                     }),
                     Event::TransferMemo(memo.into()),
                 ]